+++ sys/netiso/iso.c
@@ -4,7 +4,7 @@
  *
  * %sccs.include.redist.c%
  *
- *	@(#)iso.c	8.2 (Berkeley) 11/15/93
+ *	@(#)iso.c	8.3 (Berkeley) 01/09/95
  */
 
 /***********************************************************
@@ -387,7 +387,7 @@ caddr_t			buf;		/* RESULT: network portion of address here */
 /* ARGSUSED */
 iso_control(so, cmd, data, ifp)
 	struct socket *so;
-	int cmd;
+	u_long cmd;
 	caddr_t data;
 	register struct ifnet *ifp;
 {
